Move nodemailer from config to backend/lib/nodemailer. In paras.ts, move transporter.sendMail to after .execute. Add to do comments.
